pyproject.toml
[tool.poetry]
name = "collaboagames-backend"
version = "0.1.0"
description = "コラボゲームズバックエンドアプリケーション"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.0"
uvicorn = "^0.24.0"
sqlalchemy = "^2.0.25"
alembic = "^1.13.1"
pydantic = "^2.6.4"
psycopg2-binary = "^2.9.9"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.9"
python-dotenv = "^1.0.1"
email-validator = "^2.1.1"

[tool.poetry.dev-dependencies]
pytest = "^7.3.1"
httpx = "^0.24.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api" 
